+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [🚀] Added command support for default map style and API token update
+- [🚀] Added command support for default map style, marker color, custom marker URL and API token update
 
 ## [Next] - 2024-10-23